requirements.txt
line-bot-sdk
flask
gunicorn
pillow
pathlib